chronomancy/__init__.py -> implements new Chronomancy class
Finally have an improved implementation over `arcane_recall()` by itself!

The use of this class removes the need to explicitly use `inspect.stack()[1]`.

It does the work for you and you can think of the location of where you instantiate
the class as the "anchor_point" in the stack scope(s) where you want to to "recall"
back to.

Think Ekko from League of Legends and his Ult, it's like that now, but better.

Note: implementing this class inside the __init__.py file at the root level of the
chronomancy package because it'll be a problem if I have a `chronomancy.py` module
inside it too.

Tests included (basically just mirrors the original arcane_recall tests for now)